pyproject.toml
[tool.poetry]
name = "vishc-chatbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "vishc_chatbot"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.245"
streamlit = "^1.25.0"
streamlit-chat = "^0.1.1"
jq = "^1.4.1"
openai = "^0.27.8"
chromadb = "^0.4.3"
tiktoken = "^0.4.0"
pydantic = "1.10"
datasets = "^2.14.2"
weaviate-client = "^3.22.1"
pytest = "^7.4.0"
evaluate = "^0.4.0"
nltk = "^3.8.1"
rouge-score = "^0.1.2"
bert-score = "^0.3.13"
torch = ">=2.0.0, !=2.0.1"
sentence-transformers = "^2.2.2"
pymupdf = "^1.23.3"
seaborn = "^0.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
